More UI, basic scale information
More parts of the UI are added and there is a new data model which stores the information about the scale which shell be drawn in the WindowsFormsHost element.
Some of the elements are working now, so data can be entered and will be stored in the model. This model doesn't process the data yet.