UI is completed and painting the scale has begun
The UI is now completed, optimized and functional (it stores the data corectly in the MathModel).
The next step has been introduced: To paint the actual scale onto the WindowsFormsHost panel. Until now the base line is the only element that is painted. Its thickness depends on it's width, so the proportions won't depend on the windows size.